structure the projects
