--- react-guide/src/App.css
@@ -1,27 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.red {
-  color: red;
-}
-
-.bold {
-  font-weight: 700;
-}
-
-.redBtn {
-  background-color: red;
-}
-
-.greenBtn {
-  background-color: green;
-}
-
-.redBtn:hover {
-  background-color: orange;
-}
-
-.greenBtn:hover {
-  background-color: lightgreen;
-}
