+++ react-guide/src/containers/App.css
@@ -1,0 +1,3 @@
+.App {
+  text-align: center;
+}
