+++ react-guide/src/containers/App.test.js
